fixed missing methods in interfaces
